// src/app/pages/hr110/list/service/hr110-list-interface.ts
export interface Hr110ListInterface {
  showInformKey: { value: string };
  showPage: { value: string };
  inputText_1: { value: string };
  inputText_2: { value: string };
  inputText_3: { value: string };
  inputText_4: { value: string };
  inputText_5: { value: string };
  inputText_6: { value: string };
  inputText_7: { value: string };
  optionSelected_1: { value: string };
  optionSelected_2: { value: string };
  optionSelected_3: { value: string };
  optionSelected_4: { value: string };
  optionSelected_5: { value: string };
  optionSelected_6: { value: string };
}